refactor: rename Entries to Builtins
